Addressed 'add new user' bugs

+++ app/views/pvadmin/users/index.blade.php
@@ -16,6 +16,7 @@
         <a href="{{ URL::route('pvadmin.users.create') }}" class="btn btn-success pull-right"><span class="fa fa-plus"></span> Add New User</a>
         <h1>{{{ $title }}}</h1>
     </header>
+    <?php $allUsers = null; ?>
     @if ($links)
         {{ $users->appends(array('allUsers' => 'true'))->links() }}
     @endif
@@ -25,11 +26,19 @@
                 <li class="list-group-item">
                 <h4 class="list-group-item-heading">{{ $user->email }}</h4>
                     {{ Button::primary('Edit')->asLinkTo(route('pvadmin.users.show', ['id' => $user->id])) }}
-                    {{ Form::inline([
-                        'route' => ['pvadmin.users.destroy', $user->id],
-                        'class' => 'form-button form-button-delete',
-                        'method'    => 'delete'
-                    ]) }}
+                    @if ($links)
+                        {{ Form::inline([
+                            'route' => ['pvadmin.users.destroy', $user->id, 'allUsers' => 'true'],
+                            'class' => 'form-button form-button-delete',
+                            'method'    => 'delete'
+                        ]) }}
+                    @else
+                        {{ Form::inline([
+                                'route' => ['pvadmin.users.destroy', $user->id],
+                                'class' => 'form-button form-button-delete',
+                                'method'    => 'delete'
+                            ]) }}
+                    @endif
                         {{ Button::danger('Delete'); }}
                     {{ Form::close() }}
                 </li>
